requirements.txt
apache-airflow-providers-postgres==5.4.0
